use request timeout as default value for the queue timeout

+++ lucee-java/lucee-core/src/lucee/runtime/config/ConfigImpl.java
@@ -414,7 +414,7 @@ public abstract class ConfigImpl implements Config {
 	private Map<Key, Map<Key, Object>> tagDefaultAttributeValues;
 
 	private int queueMax=100;
-	private int queueTimeout=0;
+	private long queueTimeout=0;
 	private boolean queueEnable=false;
 	
 	
@@ -3449,10 +3449,10 @@ public abstract class ConfigImpl implements Config {
 		this.queueMax = queueMax;
 	}
 
-	public int getQueueTimeout() {
+	public long getQueueTimeout() {
 		return queueTimeout;
 	}
-	protected void setQueueTimeout(int queueTimeout) {
+	protected void setQueueTimeout(long queueTimeout) {
 		this.queueTimeout =  queueTimeout;
 	}
